Expense-Manager-Backend/requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
click==6.7
colorama==0.3.9
configparser==3.5.0
enum34==1.1.6
Flask==0.12.2
Flask-Cors==3.0.3
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.7.1
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.3.1
PyJWT==1.5.3
pylint==1.7.2
singledispatch==3.4.0.3
six==1.10.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
wrapt==1.10.11
WTForms==2.1
WTForms-JSON==0.3.3
